+++ package/stores/search.js
@@ -1,11 +1,27 @@
 import { get, readable, writable } from "svelte/store";
 import { debounce } from "../util/functional";
-export async function register(options) {
-    const { index_name, index_url, overwrite = false, wasm_url } = options;
-    const stork = window.stork;
+import { mount_script } from "../util/html";
+export async function initialize(options) {
+    const { script_url, wasm_url } = options;
+    let stork = window.stork;
+    if (script_url) {
+        if (stork) {
+            console.warn("[svelte-stork] Stork Search library was previously mounted and initialized, skipping...");
+        }
+        else {
+            try {
+                await mount_script("stork-search", script_url);
+                stork = window.stork;
+            }
+            catch (err) {
+                // TODO: Standardize error object
+                throw new ReferenceError(`bad dispatch to 'initialize' (failed to mount and initialize Stork Search library)`);
+            }
+        }
+    }
     if (!stork) {
         // TODO: Standardize error object
-        throw new ReferenceError(`bad dispatch to 'register' (Stork namespace not found)`);
+        throw new ReferenceError(`bad dispatch to 'initialize' (Stork namespace not found)`);
     }
     try {
         await stork.initialize(wasm_url);
@@ -13,7 +29,15 @@ export async function register(options) {
     catch (err) {
         // TODO: Look into what exceptions this spits out
         // and provide standardized error object
-        throw new Error("bad option 'IRegisterOptions.wasm_url' to 'register' (failed to download and initialize WASM blob)");
+        throw new Error("bad option 'IInitializeOptions.wasm_url' to 'initialize' (failed to download and initialize Stork WASM blob)");
+    }
+}
+export async function register(options) {
+    const { index_name, index_url, overwrite = false } = options;
+    const stork = window.stork;
+    if (!stork) {
+        // TODO: Standardize error object
+        throw new ReferenceError(`bad dispatch to 'register' (Stork namespace not found)`);
     }
     try {
         await stork.downloadIndex(index_name, index_url, { forceOverwrite: overwrite });