+++ src/lib/components/StorkMessage.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import type {IStorkQuery} from "../stores/search";
+    import type {IStorkResult} from "../stores/search";
 
     export let element: HTMLDivElement | null = null;
 
@@ -8,22 +8,22 @@
 
     export {_class as class};
 
-    export let query: IStorkQuery;
+    export let results: IStorkResult[];
 </script>
 
-<div bind:this={element} class="svelte-stork-message {_class}" {style}>
-    {#if query.total_hit_count > 0}
-        {query.total_hit_count} files found.
+<div bind:this={element} class="svst-message {_class}" {style}>
+    {#if results.length > 0}
+        {results.length} files found.
     {:else}
         No files found.
     {/if}
 </div>
 
 <style>
-    :global(.svelte-stork-message) {
-        padding: var(--svelte-stork-message-padding, 0.75rem 1rem);
+    :global(.svst-message) {
+        padding: var(--svst-message-padding, 0.75rem 1rem);
 
-        font-size: var(--svelte-stork-message-font-size, 2rem);
-        opacity: var(--svelte-stork-message-opacity, 0.85);
+        font-size: var(--svst-message-font-size, 2rem);
+        opacity: var(--svst-message-opacity, 0.85);
     }
 </style>